[feature]新增视频解析

+++ commonlibrary/src/main/java/com/example/commonlibrary/manager/IVideoPlayer.java
@@ -28,9 +28,9 @@ public interface IVideoPlayer {
 
     int getVolume();
 
-    int getDuration();
+    long getDuration();
 
-    int getPosition();
+    long getPosition();
 
     int getBufferedPercentage();
